Fixed all edge cases and improvised errors for usermanagement

+++ src/main/java/com/ashish/project/socialnetwork/service/UserService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 import static com.ashish.project.socialnetwork.Constants.NOT_FOUND_ERROR_CODE;
 
@@ -22,11 +22,17 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public UserDTO addUser(UserDTO userDTO) {
+    public UserDTO addUser(UserDTO userDTO) throws UserException {
+        userDTO.setId(null);
+        String email = userDTO.getEmail();
+        Optional<String> currentUseremail = userRepository.findByEmail(email);
+        if(currentUseremail.isPresent()) throw new UserException(String.format("Passed in User with user email:%s already exists in DB. Duplicate request",email), NOT_FOUND_ERROR_CODE);
         return Mapper.toDto(userRepository.save(Mapper.toEntity(userDTO)));
     }
 
     public void removeUser(Long userId) {
+        Optional<Users> currentUser = userRepository.findById(userId);
+        if(currentUser.isEmpty()) throw new UserException(String.format("Passed in User with userId:%s doesn't exists in DB Or may have already been deleted.",userId), NOT_FOUND_ERROR_CODE);
         userRepository.deleteById(userId);
     }
